Minor fixes; Notification refactor
