+++ components/FetchData.tsx
@@ -4,7 +4,7 @@ import firestore, { FirebaseFirestoreTypes } from '@react-native-firebase/firest
 import { showMessage } from "react-native-flash-message";
 import Colors from "@/constants/Colors";
 import Collections from "@/constants/Collections";
-import { SessionContext } from "@/shared/LoginSession";
+import { SessionContext } from "@/shared/Session/LoginSessionProvider";
 
 export type CommonType = {
   id: string;
@@ -171,8 +171,10 @@ const sortPatients = (patientsList: any[]) => {
 
 const fetchPatients = async () => {
   const session = useContext(SessionContext);
+  const docId = session && session.state === "valid" ?
+    session.userData.docId : undefined;
   const snapshot = await firestore().collection(Collections.Professionals)
-  .doc(session?.docId).collection(Collections.Patient).get();
+  .doc(docId).collection(Collections.Patient).get();
   const patients = snapshot.docs.map(doc => ({ id: doc.id, ...doc.data() }));
   return sortPatients(patients);
 };
@@ -181,6 +183,8 @@ export const useCheckBoxPatientsFirestoreQuery = () => {
   const queryClient = useQueryClient();
   const queryKey = ['patients'] as const;
   const session = useContext(SessionContext);
+  const docId  = session && session.state === "valid" ?
+    session.userData.docId : undefined;
 
   const { data, error, isLoading } = useQuery({
       queryKey,
@@ -190,7 +194,7 @@ export const useCheckBoxPatientsFirestoreQuery = () => {
   useEffect(() => {
     const unsubscribe = firestore()
         .collection(Collections.Professionals)
-        .doc(session?.docId)
+        .doc(docId)
         .collection(Collections.Patient)
         .onSnapshot(
             snapshot => {
@@ -221,6 +225,8 @@ export const usePatientsFirestoreQuery = () => {
   const queryClient = useQueryClient();
   const queryKey = ['patientsOfProfessional'] as const;
   const session = useContext(SessionContext);
+  const docId = session && session.state === "valid" ?
+  session.userData.docId : undefined;
 
   const { data, error, isLoading } = useQuery({
       queryKey,
@@ -230,7 +236,7 @@ export const usePatientsFirestoreQuery = () => {
   useEffect(() => {
     const unsubscribe = firestore()
         .collection(Collections.Professionals)
-        .doc(session?.docId)
+        .doc(docId)
         .collection(Collections.Patient)
         .onSnapshot(
             snapshot => {
